+++ db/migrate/20140220174630_add_abstract_and_hero_asset_to_pages_and_page_revisions.rb
@@ -1,0 +1,8 @@
+class AddAbstractAndHeroAssetToPagesAndPageRevisions < Landable::Migration
+  def change
+    add_column 'landable.pages', :abstract, :text
+    add_column 'landable.pages', :hero_asset_id, :uuid
+    add_column 'landable.page_revisions', :abstract, :text
+    add_column 'landable.page_revisions', :hero_asset_id, :uuid
+  end
+end