+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Changelog
 
 See README.md before updating this file.
+
 ## Unreleased [#](https://git.cashnetusa.com/trogdor/landable/compare/v1.5.1...master)
+* Add `categories` as a Liquid variable [#115]
 * Report full version in API responses [#117]
 
 ## v1.5.1 [#](https://git.cashnetusa.com/trogdor/landable/compare/v1.5.0...v1.5.1)